LORD IT IS LATE Media checks for users
I will need to do renaming and things after sleeping to refractor but honestly if a user comes out scott free from this check then they are a gift from the heavens. We will still not check following tweets and we don't expect them to block or unfollow kitten haters. We need to add instagram and facebook. Twitter checks their tweets, retweets, the images, their name. id. bio. profile, banner, and the last 5 for those they follow. We will probably need to track the issue causers so companies can look at that but yeah.